Deleted mgmt_systems
