+++ requirements.txt
@@ -13,6 +13,7 @@ Jinja2
 kwargify
 # 15.8.1 breaks yaycl: https://github.com/mk-fg/layered-yaml-attrdict-config/commit/ea12fbf31b96abf15543c7b436272d8854b5d324
 layered-yaml-attrdict-config
+mgmtsystem
 multimethods.py
 numpy
 ovirt-engine-sdk-python
